refactor: move dispatch functions to ContextsStates class Signed-off-by: Philippe Martin <phmartin@redhat.com>
