+++ packages/main/src/plugin/kubernetes/contexts-states.ts
@@ -26,6 +26,9 @@ import type {
 } from '/@api/kubernetes-contexts-states.js';
 import { NO_CURRENT_CONTEXT_ERROR, secondaryResources } from '/@api/kubernetes-contexts-states.js';
 
+import type { ApiSenderType } from '../api.js';
+import { dispatchTimeout } from './kubernetes-context-state-constants.js';
+
 // ContextInternalState stores informers for a kube context
 export type ContextInternalState = Map<ResourceName, Informer<KubernetesObject>>;
 
@@ -42,12 +45,64 @@ export type ContextStateResources = {
   [resourceName in ResourceName]: KubernetesObject[];
 };
 
+type ResourcesDispatchOptions = {
+  [resourceName in ResourceName]?: boolean;
+};
+
+export const dispatchAllResources: ResourcesDispatchOptions = {
+  pods: true,
+  deployments: true,
+  services: true,
+  nodes: true,
+  persistentvolumeclaims: true,
+  ingresses: true,
+  routes: true,
+  configmaps: true,
+  secrets: true,
+  // add new resources here when adding new informers
+};
+
+interface DispatchOptions {
+  // do we send information about contexts connectivity checking?
+  checkingState: boolean;
+  // do we send general context for all contexts?
+  contextsGeneralState: boolean;
+  // do we send general context for current context?
+  currentContextGeneralState: boolean;
+  // current context name
+  currentContext: string;
+  // do we send resources data for each resource kind? default false for all resources
+  resources: ResourcesDispatchOptions;
+}
+
+// Options when calling setStateAndDispatch
+interface SetStateAndDispatchOptions {
+  // true to send checking-state update event, false by default
+  checkingState?: boolean;
+  // true to send general-state update event, false by default
+  sendGeneral?: boolean;
+  // current context name
+  currentContext: string;
+  // resources for which to send a resource-specific update event
+  resources?: ResourcesDispatchOptions;
+  // the caller can modify the `previous` state in this function, which will be called before the state is dispatched
+  update: (previous: ContextState) => void;
+}
+
 export function isSecondaryResourceName(value: string): value is SecondaryResourceName {
   return secondaryResources.includes(value as SecondaryResourceName);
 }
 
 export class ContextsStates {
   private state = new Map<string, ContextState>();
+
+  private dispatchContextsGeneralStateTimer: NodeJS.Timeout | undefined;
+  private dispatchCurrentContextGeneralStateTimer: NodeJS.Timeout | undefined;
+  private dispatchCurrentContextResourceTimers = new Map<string, NodeJS.Timeout | undefined>();
+
+  private disposed = false;
+
+  constructor(private readonly apiSender: ApiSenderType) {}
 
   getContextsGeneralState(): Map<string, ContextGeneralState> {
     const result = new Map<string, ContextGeneralState>();
@@ -108,6 +163,17 @@ export class ContextsStates {
     return this.state.get(contextName)?.reachable ?? false;
   }
 
+  setStateAndDispatch(name: string, options: SetStateAndDispatchOptions): void {
+    this.safeSetState(name, options.update);
+    this.dispatch({
+      checkingState: options.checkingState ?? false,
+      contextsGeneralState: options.sendGeneral ?? false,
+      currentContextGeneralState: options.sendGeneral ?? false,
+      currentContext: options.currentContext,
+      resources: options.resources ?? {},
+    });
+  }
+
   safeSetState(name: string, update: (previous: ContextState) => void): void {
     if (!this.state.has(name)) {
       this.state.set(name, {
@@ -135,7 +201,83 @@ export class ContextsStates {
     update(val);
   }
 
-  async dispose(name: string): Promise<void> {
+  dispatch(options: DispatchOptions): void {
+    if (options.checkingState) {
+      this.dispatchCheckingState(this.getContextsCheckingState());
+    }
+    if (options.contextsGeneralState) {
+      this.dispatchGeneralState(this.getContextsGeneralState());
+    }
+    if (options.currentContextGeneralState) {
+      this.dispatchCurrentContextGeneralState(this.getCurrentContextGeneralState(options.currentContext));
+    }
+    Object.keys(options.resources).forEach(res => {
+      const resname = res as ResourceName;
+      if (options.resources[resname]) {
+        this.dispatchCurrentContextResource(resname, this.getContextResources(options.currentContext, resname));
+      }
+    });
+  }
+
+  public dispatchCheckingState(val: Map<string, CheckingState>): void {
+    this.apiSender.send(`kubernetes-contexts-checking-state-update`, val);
+  }
+
+  public dispatchGeneralState(val: Map<string, ContextGeneralState>): void {
+    this.dispatchContextsGeneralStateTimer = this.dispatchDebounce(
+      `kubernetes-contexts-general-state-update`,
+      val,
+      this.dispatchContextsGeneralStateTimer,
+      dispatchTimeout,
+    );
+  }
+
+  public dispatchCurrentContextGeneralState(val: ContextGeneralState): void {
+    this.dispatchCurrentContextGeneralStateTimer = this.dispatchDebounce(
+      `kubernetes-current-context-general-state-update`,
+      val,
+      this.dispatchCurrentContextGeneralStateTimer,
+      dispatchTimeout,
+    );
+  }
+
+  public dispatchCurrentContextResource(resname: ResourceName, val: KubernetesObject[]): void {
+    this.dispatchCurrentContextResourceTimers.set(
+      resname,
+      this.dispatchDebounce(
+        `kubernetes-current-context-${resname}-update`,
+        val,
+        this.dispatchCurrentContextResourceTimers.get(resname),
+        dispatchTimeout,
+      ),
+    );
+  }
+
+  private dispatchDebounce(
+    eventName: string,
+    value: Map<string, ContextGeneralState> | ContextGeneralState | KubernetesObject[] | Map<string, boolean>,
+    timer: NodeJS.Timeout | undefined,
+    timeout: number,
+  ): NodeJS.Timeout | undefined {
+    clearTimeout(timer);
+    if (this.disposed) {
+      return undefined;
+    }
+    return setTimeout(() => {
+      this.apiSender.send(eventName, value);
+    }, timeout);
+  }
+
+  dispose(): void {
+    this.disposed = true;
+    clearTimeout(this.dispatchContextsGeneralStateTimer);
+    clearTimeout(this.dispatchCurrentContextGeneralStateTimer);
+    for (const timer of this.dispatchCurrentContextResourceTimers.values()) {
+      clearTimeout(timer);
+    }
+  }
+
+  async disposeContext(name: string): Promise<void> {
     this.state.delete(name);
   }
