+++ packages/main/src/plugin/kubernetes/kubernetes-context-state.ts
@@ -52,21 +52,20 @@ import {
 } from '@kubernetes/client-node';
 
 import type { KubeContext } from '/@api/kubernetes-context.js';
-import type { CheckingState, ContextGeneralState, ResourceName } from '/@api/kubernetes-contexts-states.js';
+import type { ContextGeneralState, ResourceName } from '/@api/kubernetes-contexts-states.js';
 import { secondaryResources } from '/@api/kubernetes-contexts-states.js';
 import type { V1Route } from '/@api/openshift-types.js';
 
 import type { ApiSenderType } from '../api.js';
 import { Backoff } from './backoff.js';
 import { ContextsInformers } from './contexts-informers.js';
-import type { ContextInternalState, ContextState } from './contexts-states.js';
-import { ContextsStates, isSecondaryResourceName } from './contexts-states.js';
+import type { ContextInternalState } from './contexts-states.js';
+import { ContextsStates, dispatchAllResources, isSecondaryResourceName } from './contexts-states.js';
 import {
   backoffInitialValue,
   backoffJitter,
   backoffLimit,
   connectTimeout,
-  dispatchTimeout,
 } from './kubernetes-context-state-constants.js';
 import { ResourceWatchersRegistry } from './resource-watchers-registry.js';
 
@@ -97,65 +96,23 @@ interface CreateInformerOptions<T> {
   connectionDelay: NodeJS.Timeout | undefined;
 }
 
-type ResourcesDispatchOptions = {
-  [resourceName in ResourceName]?: boolean;
-};
-
-const dispatchAllResources: ResourcesDispatchOptions = {
-  pods: true,
-  deployments: true,
-  services: true,
-  nodes: true,
-  persistentvolumeclaims: true,
-  ingresses: true,
-  routes: true,
-  configmaps: true,
-  secrets: true,
-  // add new resources here when adding new informers
-};
-
-interface DispatchOptions {
-  // do we send information about contexts connectivity checking?
-  checkingState: boolean;
-  // do we send general context for all contexts?
-  contextsGeneralState: boolean;
-  // do we send general context for current context?
-  currentContextGeneralState: boolean;
-  // do we send resources data for each resource kind? default false for all resources
-  resources: ResourcesDispatchOptions;
-}
-
-// Options when calling setStateAndDispatch
-interface SetStateAndDispatchOptions {
-  // true to send checking-state update event, false by default
-  checkingState?: boolean;
-  // true to send general-state update event, false by default
-  sendGeneral?: boolean;
-  // resources for which to send a resource-specific update event
-  resources?: ResourcesDispatchOptions;
-  // the caller can modify the `previous` state in this function, which will be called before the state is dispatched
-  update: (previous: ContextState) => void;
-}
-
 // the ContextsState singleton (instantiated by the kubernetes-client singleton)
 // manages the state of the different kube contexts
 export class ContextsManager {
   private kubeConfig = new KubeConfig();
-  private states = new ContextsStates();
+  protected states: ContextsStates;
   private informers = new ContextsInformers();
   private currentContext: KubeContext | undefined;
   private secondaryWatchers = new ResourceWatchersRegistry();
-
-  private dispatchContextsGeneralStateTimer: NodeJS.Timeout | undefined;
-  private dispatchCurrentContextGeneralStateTimer: NodeJS.Timeout | undefined;
-  private dispatchCurrentContextResourceTimers = new Map<string, NodeJS.Timeout | undefined>();
 
   private connectTimers = new Map<ResourceName, NodeJS.Timeout | undefined>();
   private connectionDelayTimers = new Map<string, NodeJS.Timeout | undefined>();
 
   private disposed = false;
 
-  constructor(private readonly apiSender: ApiSenderType) {}
+  constructor(readonly apiSender: ApiSenderType) {
+    this.states = new ContextsStates(apiSender);
+  }
 
   setConnectionTimers(
     resourceName: ResourceName,
@@ -269,7 +226,7 @@ export class ContextsManager {
         (contextChanged && name === this.currentContext?.name) ||
         (checkOnlyCurrentContext && name !== this.currentContext?.name)
       ) {
-        await this.states.dispose(name);
+        await this.states.disposeContext(name);
         await this.informers.dispose(name);
         removed = true;
       }
@@ -315,7 +272,7 @@ export class ContextsManager {
       for (const context of this.kubeConfig.contexts) {
         if (added.includes(context.name)) continue;
         if (!this.compareContexts(context.name, this.kubeConfig, previousKubeConfig)) {
-          await this.states.dispose(context.name);
+          await this.states.disposeContext(context.name);
           await this.informers.dispose(context.name);
           const kubeContext: KubeContext = this.getKubeContext(context);
           const informers = this.createKubeContextInformers(kubeContext);
@@ -325,10 +282,11 @@ export class ContextsManager {
     }
 
     if (added.length || removed || contextChanged) {
-      this.dispatch({
+      this.states.dispatch({
         checkingState: false,
         contextsGeneralState: true,
         currentContextGeneralState: true,
+        currentContext: this.kubeConfig.currentContext,
         resources: dispatchAllResources,
       });
     }
@@ -414,8 +372,9 @@ export class ContextsManager {
     let connectionDelay: NodeJS.Timeout | undefined;
     this.setConnectionTimers('pods', timer, connectionDelay);
 
-    this.setStateAndDispatch(context.name, {
+    this.states.setStateAndDispatch(context.name, {
       checkingState: true,
+      currentContext: this.kubeConfig.currentContext,
       update: state => {
         state.checking = { state: 'checking' };
       },
@@ -428,15 +387,17 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: { pods: true },
           update: state => state.resources.pods.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: { pods: true },
           update: state => {
             state.resources.pods = state.resources.pods.filter(o => o.metadata?.uid !== obj.metadata?.uid);
@@ -445,17 +406,19 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: { pods: true },
           update: state =>
             (state.resources.pods = state.resources.pods.filter(d => d.metadata?.uid !== obj.metadata?.uid)),
         });
       },
       onReachable: reachable => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           checkingState: true,
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: dispatchAllResources,
           update: state => {
             state.checking = { state: 'waiting' };
@@ -465,9 +428,10 @@ export class ContextsManager {
         });
       },
       onConnectionError: error => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           checkingState: true,
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: dispatchAllResources,
           update: state => {
             state.checking = { state: 'waiting' };
@@ -491,15 +455,17 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: { deployments: true },
           update: state => state.resources.deployments.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: { deployments: true },
           update: state => {
             state.resources.deployments = state.resources.deployments.filter(
@@ -510,8 +476,9 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           sendGeneral: true,
+          currentContext: this.kubeConfig.currentContext,
           resources: { deployments: true },
           update: state =>
             (state.resources.deployments = state.resources.deployments.filter(
@@ -535,13 +502,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { configmaps: true },
           update: state => state.resources.configmaps.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { configmaps: true },
           update: state => {
             state.resources.configmaps = state.resources.configmaps.filter(o => o.metadata?.uid !== obj.metadata?.uid);
@@ -550,7 +519,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { configmaps: true },
           update: state =>
             (state.resources.configmaps = state.resources.configmaps.filter(
@@ -574,13 +544,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
           resources: { secrets: true },
+          currentContext: this.kubeConfig.currentContext,
           update: state => state.resources.secrets.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { secrets: true },
           update: state => {
             state.resources.secrets = state.resources.secrets.filter(o => o.metadata?.uid !== obj.metadata?.uid);
@@ -589,7 +561,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { secrets: true },
           update: state =>
             (state.resources.secrets = state.resources.secrets.filter(d => d.metadata?.uid !== obj.metadata?.uid)),
@@ -616,13 +589,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { persistentvolumeclaims: true },
           update: state => state.resources.persistentvolumeclaims.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { persistentvolumeclaims: true },
           update: state => {
             state.resources.persistentvolumeclaims = state.resources.persistentvolumeclaims.filter(
@@ -633,7 +608,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { persistentvolumeclaims: true },
           update: state =>
             (state.resources.persistentvolumeclaims = state.resources.persistentvolumeclaims.filter(
@@ -657,13 +633,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { nodes: true },
           update: state => state.resources.nodes.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { nodes: true },
           update: state => {
             state.resources.nodes = state.resources.nodes.filter(o => o.metadata?.uid !== obj.metadata?.uid);
@@ -672,7 +650,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { nodes: true },
           update: state =>
             (state.resources.nodes = state.resources.nodes.filter(d => d.metadata?.uid !== obj.metadata?.uid)),
@@ -694,13 +673,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { services: true },
           update: state => state.resources.services.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { services: true },
           update: state => {
             state.resources.services = state.resources.services.filter(o => o.metadata?.uid !== obj.metadata?.uid);
@@ -709,7 +690,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { services: true },
           update: state =>
             (state.resources.services = state.resources.services.filter(d => d.metadata?.uid !== obj.metadata?.uid)),
@@ -731,13 +713,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { ingresses: true },
           update: state => state.resources.ingresses.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { ingresses: true },
           update: state => {
             state.resources.ingresses = state.resources.ingresses.filter(o => o.metadata?.uid !== obj.metadata?.uid);
@@ -746,7 +730,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { ingresses: true },
           update: state =>
             (state.resources.ingresses = state.resources.ingresses.filter(d => d.metadata?.uid !== obj.metadata?.uid)),
@@ -774,13 +759,15 @@ export class ContextsManager {
       backoff: new Backoff(backoffInitialValue, backoffLimit, backoffJitter),
       connectionDelay: connectionDelay,
       onAdd: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { routes: true },
           update: state => state.resources.routes.push(obj),
         });
       },
       onUpdate: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { routes: true },
           update: state => {
             state.resources.routes = state.resources.routes.filter(
@@ -791,7 +778,8 @@ export class ContextsManager {
         });
       },
       onDelete: obj => {
-        this.setStateAndDispatch(context.name, {
+        this.states.setStateAndDispatch(context.name, {
+          currentContext: this.kubeConfig.currentContext,
           resources: { routes: true },
           update: state =>
             (state.resources.routes = state.resources.routes.filter(
@@ -899,8 +887,9 @@ export class ContextsManager {
     options: CreateInformerOptions<T>,
   ): void {
     if (options.checkReachable) {
-      this.setStateAndDispatch(context.name, {
+      this.states.setStateAndDispatch(context.name, {
         checkingState: true,
+        currentContext: this.kubeConfig.currentContext,
         update: state => {
           state.checking = { state: 'checking' };
         },
@@ -932,88 +921,6 @@ export class ContextsManager {
         }
       }, nextTimeout);
     });
-  }
-
-  private setStateAndDispatch(name: string, options: SetStateAndDispatchOptions): void {
-    this.states.safeSetState(name, options.update);
-    this.dispatch({
-      checkingState: options.checkingState ?? false,
-      contextsGeneralState: options.sendGeneral ?? false,
-      currentContextGeneralState: options.sendGeneral ?? false,
-      resources: options.resources ?? {},
-    });
-  }
-
-  private dispatch(options: DispatchOptions): void {
-    if (options.checkingState) {
-      this.dispatchCheckingState(this.states.getContextsCheckingState());
-    }
-    if (options.contextsGeneralState) {
-      this.dispatchGeneralState(this.states.getContextsGeneralState());
-    }
-    if (options.currentContextGeneralState) {
-      this.dispatchCurrentContextGeneralState(
-        this.states.getCurrentContextGeneralState(this.kubeConfig.currentContext),
-      );
-    }
-    Object.keys(options.resources).forEach(res => {
-      const resname = res as ResourceName;
-      if (options.resources[resname]) {
-        this.dispatchCurrentContextResource(
-          resname,
-          this.states.getContextResources(this.kubeConfig.currentContext, resname),
-        );
-      }
-    });
-  }
-
-  public dispatchCheckingState(val: Map<string, CheckingState>): void {
-    this.apiSender.send(`kubernetes-contexts-checking-state-update`, val);
-  }
-
-  public dispatchGeneralState(val: Map<string, ContextGeneralState>): void {
-    this.dispatchContextsGeneralStateTimer = this.dispatchDebounce(
-      `kubernetes-contexts-general-state-update`,
-      val,
-      this.dispatchContextsGeneralStateTimer,
-      dispatchTimeout,
-    );
-  }
-
-  public dispatchCurrentContextGeneralState(val: ContextGeneralState): void {
-    this.dispatchCurrentContextGeneralStateTimer = this.dispatchDebounce(
-      `kubernetes-current-context-general-state-update`,
-      val,
-      this.dispatchCurrentContextGeneralStateTimer,
-      dispatchTimeout,
-    );
-  }
-
-  public dispatchCurrentContextResource(resname: ResourceName, val: KubernetesObject[]): void {
-    this.dispatchCurrentContextResourceTimers.set(
-      resname,
-      this.dispatchDebounce(
-        `kubernetes-current-context-${resname}-update`,
-        val,
-        this.dispatchCurrentContextResourceTimers.get(resname),
-        dispatchTimeout,
-      ),
-    );
-  }
-
-  private dispatchDebounce(
-    eventName: string,
-    value: Map<string, ContextGeneralState> | ContextGeneralState | KubernetesObject[] | Map<string, boolean>,
-    timer: NodeJS.Timeout | undefined,
-    timeout: number,
-  ): NodeJS.Timeout | undefined {
-    clearTimeout(timer);
-    if (this.disposed) {
-      return undefined;
-    }
-    return setTimeout(() => {
-      this.apiSender.send(eventName, value);
-    }, timeout);
   }
 
   public getContextsGeneralState(): Map<string, ContextGeneralState> {
@@ -1058,11 +965,7 @@ export class ContextsManager {
 
   public dispose(): void {
     this.disposed = true;
-    clearTimeout(this.dispatchContextsGeneralStateTimer);
-    clearTimeout(this.dispatchCurrentContextGeneralStateTimer);
-    for (const timer of this.dispatchCurrentContextResourceTimers.values()) {
-      clearTimeout(timer);
-    }
+    this.states.dispose();
     for (const timer of this.connectTimers.values()) {
       clearTimeout(timer);
     }
